refactor(precompiles): remove superseded B-20 capability traits
The `common::ops` capability traits predate the SOLID versioned-logic
refactor. Every operation they provided — transfer, mint, burn, pause,
role management, permit — is now implemented independently by each
variant's `logic/interface.rs` + `logic/vN`, so the traits had no
production callers left: nothing outside `common/` referenced them, and
the only implementor was a `TestToken` fixture that existed solely to
keep their own unit tests compiling.

Removes `Transferable`, `Mintable`, `Burnable`, `Pausable`,
`Configurable`, `Permittable`, and `RoleManaged`, along with `TestToken`.

Kept, since these are live shared surface:
- `B20Guards` — authorization/policy guards used by every version
- `B20TokenRole` — built-in role ids, plus the tests pinning them to
  base-std's keccak strings
- `Eip712Domain` / `PermitArgs` — used by the stablecoin and asset
  permit logic. `permittable.rs` is renamed to `permit.rs` now that only
  the argument type remains.

`B20Guards`' tests move to the existing `TestStablecoinToken` fixture
rather than a bespoke one.

`domain_typehash_matches_eip712_domain_type` was the only check pinning
`DOMAIN_TYPEHASH` to its EIP-712 type string, and each `logic/vN` keeps
its own frozen copy of that constant. Rather than lose the check, it now
lives in all four version test modules, matching the existing convention
that a frozen version's coverage must not depend on shared scaffolding.

No behavior change: all 843 precompile tests pass, golden state roots
unchanged.
