+++ crates/common/precompiles/src/common/mod.rs
@@ -7,10 +7,7 @@ mod core_storage;
 pub use core_storage::B20CoreStorage;
 
 mod ops;
-pub use ops::{
-    B20Guards, B20TokenRole, Burnable, Configurable, Eip712Domain, Mintable, Pausable, PermitArgs,
-    Permittable, RoleManaged, Transferable,
-};
+pub use ops::{B20Guards, B20TokenRole, Eip712Domain, PermitArgs};
 
 mod pausable_feature;
 pub use pausable_feature::B20PausableFeature;
@@ -21,9 +18,7 @@ pub use policy_type::B20PolicyType;
 #[cfg(any(test, feature = "test-utils"))]
 pub(super) mod test_utils;
 #[cfg(any(test, feature = "test-utils"))]
-pub use test_utils::{
-    FakePolicyAccounting, InMemoryTokenAccounting, TestStablecoinToken, TestToken,
-};
+pub use test_utils::{FakePolicyAccounting, InMemoryTokenAccounting, TestStablecoinToken};
 
 mod token;
 pub use token::Token;
